arm: dts: pxa1U88: update sdh dts for rx tuning
Add some new perpority to 3 sd hosts for support rx tuning.

Now eMMC HS200 is enabled.

As there is still some risk to enable SDR104 mode for sd card and sdio
device. But we still update the relate setting in dts, in case we want
to debug and test it.

Change-Id: Ie64a0202913cf07f82f760e36e43d24f0758b38e
Signed-off-by: Jialing Fu <jlfu@marvell.com>