arm64: dts: pxa1908: update sdh dts for rx tuning
Add some new perpority to 3 sd hosts for support rx tuning.

Now eMMC HS200 is enabled.

As there is still some risk to enable SDR104 mode for sd card and sdio
device. But we still update the relate setting in dts, in case we want
to debug and test it.

Change-Id: Ia03b7c3aa398cf21e69ebacf8d7a7d48609f1611
Signed-off-by: Jialing Fu <jlfu@marvell.com>