Makefile: drivers: compile mmc later than clk
In mmc driver, it has clock setting operations, so it is better to
compile clk driver earlier.

Currently, we just enable SDH Rx tuning with DVFS enabled in pxa platfrom.
But from the real test, we found DVFS function is not ready for SDH driver.
And this is a fatal issue need to be fixed.

Considering to less impact to current code base, we just let mmc driver
to be compiled later.

Change-Id: I76d890bd07d33dbdd4c8379b69314f8e0919aa3a
Signed-off-by: Jialing Fu <jlfu@marvell.com>