marvell: telephony: msocket: reset CP if CP is synced when reboot
when AP reset, AP have to clean the resources that ocuppied by CP and then hold CP.
acipc_reset_cp_request have to send a message to CP let CP know that AP will hold CP,
after received CP response for this request, AP can hold cp and free the resources that
ocuppied by CP.
if CP have not synced with AP or CP is down, it can not make a response for the request,
so AP may have to wait until timeout, this wait is not approprite. so this patch judge
if CP have not synced with AP or CP is down, it is no need to call acipc_reset_cp_request
to CP since the data path is not working.

Change-Id: Ic334596197e510e383600d276b851333fa7dae8d
Signed-off-by: Nan Xiang <nanxiang@marvell.com>