clk: audio-map: refine audio power on/off for ulc1
refine audio island power on/off func for ulc1. Bit 1 of APMU audio
clock register is used to reset SSPA & ADMA when setting to 1, so
we should not configure it to 1 for normal use. DE also suggests
it doesn't impact if SSPA & ADMA are not in reset, when audio is in
power-off state. So the final solution is that we don't touch this
bit on ULC1.

This bit has different meaning comparing to Helan2 & Eden, so we
add a new function for ULC1.

Change-Id: I047c2765eb4cc2eb2c245f3fc3cfb8842cb45eb1
Signed-off-by: Qiao Zhou <zhouqiao@marvell.com>