power: 88pm886: fix automatic recharge
we disable the automatic recharge by clearing
CHG_ENABLE bit on end-of-charger event,
and restart charging by SW when voltage goes below
the desired threshold.
however, if the threshold is too close to the target voltage,
there might be a continuous loop of start-stop charging.
to solve it, set the auto-recharge level to minimum (50mV),
and use the recharge-threshold from DTS in order to restart
the charging by SW.

Change-Id: I0006d7d50d58d97cc96a3a066899746aa1207e4f
Signed-off-by: Shay Pathov <shayp@marvell.com>