power: 88pm886: fix over voltage fault
on A0 revision there is a faulty OV_VBAT when charging starts.
in order to overcome it, the solution is to disable
OV_VBAT_EN, OV_VSYS1 and UV_VSYS1.
on dead battery case, this will not be enough because PMIC is reset,
and these are re-enabled by default,
so also need to change the default value of them.

Change-Id: I5900cbb78db06ca32c45d85e7c1ed845f6de3649
Signed-off-by: Shay Pathov <shayp@marvell.com>