cpufreq: mmp: support multi-cluster cpufreq
On platform with multi-cluster, different cluster
may use different policy. In each cluster, the
first cpu is selected to manage the policy for
the whole cluster. In cpufreq target callback,
we will not set rate directly, but update request
to cpufreq_min qos. cpufreq_min qos works as a
client for cpufreq_max qos and the final target
rate is set in cpufreq_max qos notifier. In this
version, each cluster can change its rate freely.

Change-Id: I2baa5c90c8586a1cc4333abea52ad4b4679f23cb
Signed-off-by: Bill Zhou <zhoub@marvell.com>