dma: mmp-tdma: refine tdma update pointer
In current TDMA, the tdma active status and source/destination
registers in chain mode are not updated until tdma client sends
a request. It's a HW behavior.

It causes one issue that when audio stream starts up, the pointer
which is from source/destination register is not the actual value.
When ASoC framework uses incorrect pointer to copy data from/to
user space, it causes wrong data at the head.

So now, add some local status to mark whether tdma start status
when the registers are not updated. We can't use the dma status
in dma engine framework, and we can't modify freely to our purpose
either.

Change-Id: Id75f30a0a637f9c7a93148c0458fa44471c784b0
Signed-off-by: Qiao Zhou <zhouqiao@marvell.com>