power: 88pm886: fix charging stop
When a charger provides a high VBUS and the charging current is high,
disabling the charge by clearing CHG_ENABLE bit can fail.
Disabling the charge process will cause a sudden drop on the VBUS current.
this drop will generated an inductive ringing on VBUS that causes a charge removal detection.
to overcome this issue,
the charge must be stopped by setting the USB_SUSP bit.
once the charge is stopped, both CHG_ENABLE and USB_SUSP bits can be cleared

Change-Id: I5cabbd278c46382d7f78c481e6fa29d4c508dbc4
Signed-off-by: Shay Pathov <shayp@marvell.com>