ASoC: mmp-sspa: remove audio mode settings
This is to refine the audio mode feature for audio MAP across
platforms. Audio mode control in SSPA driver can't cover the FM
radio user scenario, since it doesn't use SSPA driver. So we move
the controls in mmp-map-fe driver, which has a better view whether
audio MAP needs to enable this feature or not.

Change-Id: Ia5e5fc990e1fa65aec42521dcc04d8f24c6e1796
Signed-off-by: Qiao Zhou <zhouqiao@marvell.com>