media: B52: splite IDI subdev into two
For B52ISP v3.2.4, it has two input and two output ports in IDI,
each of the input can be routed to each output. If use only one
IDI to represent the IDI, there is no way to figure out how data
flow is configured from outside. The solution is to split the
IDI subdev into two, so that data flow can be represented by
media-link.
1> add IDI port number and dump_enable flag in ISP H/W descriptor
2> add H/W descriptor for B52ISP-Lite
3> change the IDI related code to support per port operation
4> fix potential bug in data stream topology discover

Change-Id: Ib747792407ef9d8df5fff9d59b761e221a57d13c
Signed-off-by: Jiaquan Su <jqsu@marvell.com>
Signed-off-by: Owen Zhang <xinzha@marvell.com>