ASoC: soc-pcm: check user when do BE hw-free
This patch is to resolve CQ00069453.

When two FEs use one BE, in some case after BE
is trigger stop, it can’t be trigger start any
more. Root cause is:
1. FE(I2S4) try to trigger stop BE when stop HiFi,
   but not success since FE(I2S1) still use it.
2. FE(I2S1) try to trigger stop BE when VoIP xrun,
   BE is trigger stop since no FE use it.
3. Up layer close HiFi DAPM path, BE is in hw-free
   state.
4. Since VoIP will not close and reopen DAPM path,
   BE state is hw-free. When write data and try to
   trigger start BE, since BE state is wrong, so
   BE is not really trigger start.

To resolve this issue, add check user when do BE
hw-free. If its user is more than one, don't do
hw-free and shut down.

Change-Id: I51364f9e8898ebffa8d42f51dd3573e70e495214
Signed-off-by: Leilei Shang <shangll@marvell.com>