ARM64: disable preemption in machine_shutdown
Port from one patch for arch/arm/kernel/process.c.

Since the smp call to stop the other cpus are handled in those
cpus in interrupt context, there's a potential for those smp
handlers to interrupt threads holding spin locks (such as the
one a mutex holds).  This prevents those threads from ever
releasing their spin lock, so if the cpu doing the shutdown
is allowed to switch to another thread that tries to grab the
same lock/mutex, we could get into a deadlock (the spin lock
call is called with preemption disabled in the mutex lock code).

To avoid that possibility, disable preemption before doing the
smp_send_stop().

Change-Id: I4e360654747dd96f1b612d31637a5d19348e07e3
Signed-off-by: Raul Xiong <xjian@marvell.com>