clk: mmp: dual clusters use different bridge op on Helan3
Bridge op is used when we change frequencies
on the same clock source on PLL2, PLL3 or
PLL4. Currently PLL1_1248M is used for dual
clusters, but the bridge op is different.
For example, (pclk:1248, core_aclk:624) is
used for dual clusters, but other fields of
this pp such as L1/L2 rtcwtc value/address
are different.

Change-Id: I2c7d96d05a698fd2f2db8084b699883cfefa226b
Signed-off-by: Bill Zhou <zhoub@marvell.com>