drivers: battery: disable auto temperature monitor
for 88pm886 A1 chip and forward there's a new mechanism
to allow battery temperature automatic HW control so
charging will be performed only when battery temperature
is in a valid range.

when charger driver is up this mechanism should be disabled
and the control for charging & temperature monitoring is
transferred to the driver.

Change-Id: I3dc4b17298e9ac9d16ad59c192f7a8e36622100c
Signed-off-by: Amit Sharon <asharon@marvell.com>