leds: 88pm886: optimize cf-gpio usage
This gpio isn't shared with other peripherals, so instead of request it
over and over again, we will require it in probe function, and release
when the both led devices will be removed.
For implement that replacement of devm_gpio_xx -> gpio_xx was required,
and defined gpio_en only as static variant.

Change-Id: I1b2164adcf6a55f9acb4aebb58954a8ff5a2df77
Signed-off-by: Ehud Naim <ehudn@marvell.com>