arm64: mcpm: l2 cache is only flushed by last man in suspend
Now L2 cache is only flushed in below conditions:
1. The cpu is the last man;
2. The power off cpu action is called in suspend scenario;

We made this patch because in ULC, FM radio can only enter D1 in
suspend, which is also used in idle. The difference is in suspend
D1, l2 cache need to be flushed while in idle D1 l2 cache is kept.
This patch can distinguish the 2 scenarios.

Change-Id: I29075afc0dd01cb391c09737bf1ae2cb86c56d20
Signed-off-by: Fangsuo Wu <fswu@marvell.com>