power: 88pm886: extend recharge condition
after reaching EOC, the charger driver monitors the VBAT
and restarts the charging when it crosses a threshold.
we would like to ignore momentary drops on VBAT in this condition,
so change the recharge monitor to read VBAT 3 times instead of 1.
also change the state machine usage to workqueue,
in order to avoid using delays inside threaded irq context

Change-Id: I77f6ed02e45db36426a1b810e1f3534e0369b5c7
Signed-off-by: Shay Pathov <shayp@marvell.com>