clk: mmp: skip default rate check on PLL3 on Helan3
Output of PLL3 VCO -- PLL3p provides
multiple op for cluster-1. If the boot
up op set in uboot is not same as the
default rate statically defined in kernel
side, a panic will be triggered. This patch
appends a flag to vco clock in order to
skip this type of check.

Change-Id: I6fa24d3b838fbcde53ba0bb9b9d47e3f500dc40b
Signed-off-by: Bill Zhou <zhoub@marvell.com>