power: 88pm886: disable auto recharge limitation
in order for PM886 to start charging, it is required that
VBAT will be lower than the recharge threshold for at least 128ms.
this might cause a problem when SW tries to enable charging,
but the charging doesn't really starts.
to overcome this limitation,
we override the status of the internal comparator
after the 128ms filter, and then enable charging.
after charging is enabled, we reset the values to the original configuration.

Change-Id: I56f4e49d409bb546a9caa01efbf4591996e101c2
Signed-off-by: Shay Pathov <shayp@marvell.com>