power: 88pm88x: change voltage notification for recharge
battery driver sends a notification when VBAT is changed
in 100mV since the last notification.
the purpose is that we are using SW recharge, and the charger driver
is waiting for this notification in order to check the voltage
and decide if to start recharge or not.
There can be cases that charger driver is waiting,
but the notification will come too late.

A better solution is to remove the voltage check,
and notify whenever the status is FULL.

Change-Id: I123359ba1da2b81a1217087ce92d8a6e39c49dd7
Signed-off-by: Shay Pathov <shayp@marvell.com>