clk: mmp: fix logic issue for cpu rtc/wtc
We update cpu l1/l2 rtc/wtc only when there
is real core FC and l1 xtc of target op is
different from that of current op. But we
have a case that l1 xtc is same but l2 is
different. Under this condition, l2 xtc
settings will not be changed.

Change-Id: I311dc742c1777a2d9a2e6ff31201b1328b5e5be2
Signed-off-by: Bill Zhou <zhoub@marvell.com>