mfd: 88pm886: fix override of RTC_ALARM_CTRL1 register
Using regmap_register_patch() overrides the entire value of the register,
which causes RTC alarm interrupt to be missed in charging mode. in that
case, we just want to enable USE_XO bit, so replace it with read-modify-
write. As result of re-organization of 88pm886, part of this fix was lost.

based on:
http://shgit.marvell.com/r/#/c/136169/4

Change-Id: I968f5a527a5c9c77b8ebcde24849dcea6f66c099
Signed-off-by: Ehud Naim <ehudn@marvell.com>