clk: mmp: fix logic issue for cpu rtc/wtc in ULC1
We update cpu l1/l2 rtc/wtc only when there is real core FC
and l1 xtc of target op is different from that of current op.
But we have a case that l1 xtc is same but l2 is different.
Under this condition, l2 xtc settings will not be changed.

Change-Id: Iecf1627000a7752d7ef6a232e50b42207b072301
Signed-off-by: Fiona Yang <chy@marvell.com>