configs: enable USB_GADGET_CHARGE_ONLY for pxa1936
Enable USB_GADGET_CHARGE_ONLY for pxa1936, with charge only mode,
D+/D- is disabled and there is no data transferred on USB bus.
Charge only mode is good for data safety and can let system suspend
when connected to PC, so that chagering speed can be a little faster.

Change-Id: If7b61795df448d84dff8e24f89e18d3b712d8c89
Signed-off-by: Leo Song <liangs@marvell.com>