power: 88pm88x_charger: fix power event report
In the state machine of the charger, we notify events with
power supply information.
since charger mode application refer to the information of the
reported power supply type, need to report with the appropriate
one.
add new parameter to hold used power supply type, and report
according to this parameter.

Signed-off-by: Yaacov Hazan <yaacovh@marvell.com>
Change-Id: Ibb18b82348cd74862d00b4c47d963503530dd678