mmc: sdhci: sdhci_ops: add pm runtime for callbacks
In sdhci-pxav3.c, we try to disable sdh's function/axi clock when
sdh is in runtime suspended, this is helpful to save power.

But we need to follow a strict rule:
The callbacks exported by sdhci_ops should be protected by
pm_runtime_get/put. Making sure AXI clock is enabled when SDH registers
are being accessed.

Note:
1) For sdhci_pre_req and sdhci_post_req, there is not pm_runtime_get/put.
But it is safe. As there two functions only handler the DMA operations
without access SDH registers.
2) Other APIs had been protected by pm_runtime_get/put before.

Change-Id: Ic52afe03c43e108a8f25d3fc5465e7c7ed582381
Signed-off-by: Jialing Fu <jlfu@marvell.com>