arm64: add Cortex-A53 cache errata workaround
It is based on the patch from linux-next

arm64: add Cortex-A53 cache errata workaround

The ARM errata 819472, 826319, 827319 and 824069 define the same
workaround for these hardware issues in certain Cortex-A53 parts.
Use the new alternatives framework and the CPU MIDR detection to
patch "cache clean" into "cache clean and invalidate" instructions if
an affected CPU is detected at runtime.

The patch is commited by Andre Przywara <andre.przywara@arm.com>
and Will Deacon <will.deacon@arm.com>

The original patch is based on the errata framework set up by ARM.
The patches for the framework are submitted after 3.14, and we do not
want to port all the patches. So just pick up the fix, and use the
old way to do it.

The erreta 819472 are only for r0p1, and our earliest chip is r0p2.
So we do not need define the erreta 819472.
For other erreta 826319, 827319 and 824069, they are fixed from r0p3.
So we still need these erretaes.

Change-Id: Ide06417a162fac7f2d2fb92ac7597a49a079448f
Signed-off-by: Chao Xie <chao.xie@marvell.com>