video: mmp: remove __clk_lookup operation in dfc
As __clk_lookup will iterate all children clk of a parent clk, some
children clk are no needed to be cared, if one child clk is deleting
while __clk_lookup is checking it, it will cause panic about clk_lookup.
So remove __clk_lookup operation when do dfc, only __clk_lookup once in
init, and save the clk node, which will be used when do dfc. It is safe
and more effective.

Change-Id: Iceba3d3b066dc348c4d67dacd8bfa0550c966240
Signed-off-by: Qing Zhu <qzhu@marvell.com>