input: 88pm88x_onkey: fix initialization order
currently in the onkey driver we're calling
the init of the device tree at the beginning of the probe
before the input device is being initialized.
however there's a print in the device tree initialization
which relies on the input device.

this patch will move the init of the device tree after the
input device has finished to initialized

Change-Id: I435acf0be0bb5e489f4561ebb2f837cc70fbd09d
Signed-off-by: Amit Sharon <asharon@marvell.com>