clk: extend the usage of clk flag CLK_SET_RATE_GATE
Clk flag CLK_SET_RATE_GATE is expected to be used for clk whose
rate change must be issued when it is gated.
Previously it only uses prepare_count to identify if the clock
is enabled or not. But sometimes, the clock is enabled by bootloader
instead of kernel, in this case the flag won't take effect.
Extend its usage, add __clk_is_enabled as one extra condition to
check if clock is enabled or not. Clk driver responsible to provide
is_enabled callback to show the clock's real status.

Change-Id: I5b5e3f5d3974beec67d1ee8147b8478fe920e031
Signed-off-by: Zhoujie Wu <zjwu@marvell.com>