cpufreq: mmp_bL: skip to set clock rate when all cores in a cluster are offline
When all cores in a cluster are offline, the
policy structure will be freed. In this case,
we don't set clock rate. Though currently
cpu hotplug only happens during suspend and
resume, it's better to add this protection
for hotplug all cores in a cluster by
"echo 0 > /sys/devices/system/cpu/cpux/online".
When at least a core in that cluster becomes
online, the policy and associated structures
will be setup again in cpufreq init callback.

Change-Id: Ieccd812484e4ce04b1a4e7b89d4a203b013db4e7
Signed-off-by: Bill Zhou <zhoub@marvell.com>