arm64: dts: set mmc1_dat[4-7] as mmc1_dat
set the mmc_dat[4-7] mfpr pins as alternate function 0 (AF0).
if set as function1 (gpio), it creates a contention, since these
gpios are used on other interfaces.
for example: gpio20 is the proximity sensor interrupt.

although not used by mmc1, setting them as mmc1 is the best
option since they are unused by this interface and do not have any
contention with other interfaces

Change-Id: I2f47c4d45b3ea74a5f309c1047bb70964d189201
Signed-off-by: Yoav Gvili <ygvili@marvell.com>