media: soc_camera: reset SC2 before stream on
As DE suggested, it need disable SC2 all clks when stream off,
and enable SC2 all clks when stream on, it fixed the following issue:
the user do stream off but the driver SC2 DMA hasn't receive DMA EOF,
it will cause the DMA engine in no-done status, it will impact the next
stream on.
do reset SC2, it can keep the DMA engine in clean status before stream on.

Change-Id: Ifa4de593d4a4aa8bac7135be6c2d4b062ec0c3d8
Signed-off-by: Owen Zhang <xinzha@marvell.com>