power: 88pm88x_charger: delay after setting CHG_EN bit
to start charging in 88pm880/88pm886,
VBAT must be lower than the recharge threshold for at least 128ms.
this can causes a potential issue that the charging fails to start
even the software set the enable bit.

to fix it:
* override the status of internal comparator after the 128ms filter ends
* enable charging
* then reset to use original configuration

as recommended by Si design team,
add a delay which is at least 3ms after CHG_EN is set,
before resetting the values to the original configurations

Change-Id: Ia13feeb2dd7e48102d43b08b1169eee829956e52
Signed-off-by: Shay Pathov <shayp@marvell.com>