mmc: sdhci: set vqmmc to 3V3 and 1V8 accurately
In previous code:
2.7V~3.6V is passed to regulator_set_voltage when MMC_SIGNAL_VOLTAGE_330
and 1.7V~1.95V is passed when MMC_SIGNAL_VOLTAGE_180.

Although this make sure the voltage can be set successfully by regulator,
it cause the voltage may have big gap to 3.3V or 1.8V, and some SD card
can't work well as CRC error.

So hack it to set vqmmc to 3V3 or 1V8 accurately. But if using new PMIC
chip, please check whether the voltage can be set successfully.

Change-Id: Ie085d35fb3ca25881ba74e19c641de5ce9f52a08
Signed-off-by: Jialing Fu <jlfu@marvell.com>