arm64: dts: block D2 when MAP in use for Helan3
Quote from spec:
The audio subsystem does not have its own power domain.
It is in the Utra Drowsy Register power domain, which
is same as the Application subsystem. The audio
subsystem can only operate in D0 and D1 state.

This patch is to fix CQ00084647. If system goes to D2
when FM is working, audio subsystem can't work, so mute.

Change-Id: I7d033bd3b71235cd211c8607887674ba29e99077
Signed-off-by: Leilei Shang <shangll@marvell.com>