arm64: dts: add dts properties for lm3532
This patch adds 2 more properties to the lm3532 driver in order to allow
more flexible and generic handling in the driver.
This patch also modifies vendor name in "compatible" property to be "ti".
In addition the dts replace old numeric constants with more descriptive
definition in the driver's properties.
The properties added are
1. feedback_en_val: configures which current sinks are part of the boost
   control loop.
2. output_cfg_val: configures how the three control banks are routed to
   the current sinks (ILED1, ILED2, ILED3).

Change-Id: Id01cac7405cca5c0fa6aa28855818d43e93b7385
Signed-off-by: Ohad Sharabi <sohad@marvell.com>