power: 88pm88x: support ir drop compensation
When a battery is charged, some contributing factors
increase the VBAT voltage sensed by PMIC. These factors
are the PCB and battery connector resistance, the battery
internal resistance and the ground resistance. The
weight of their effect is proportional to the charging
current; the higher the current, the greater will be the
difference between VBAT and VBAT_INT. The difference
between VBAT and VBAT_INT will become negligible as the
battery reaches the target voltage and the current decrease.
The final result is prolonged charging time.
To overcome this effect, a compensation mechanism has been
added in 88pm880. A value of the battery internal resistance
must be programming to the device by software, and the
device computes the voltage drop by multiplying the battery
current measured by the Sigma-Delta by the resistance value.
The resulting voltage is added to the programmed
fast charge terminal voltage to compensation for the IR drop.

Change-Id: I49320d2a094d72119cc56981568f3b8f245fbb81
Signed-off-by: Fenghang Yin <yinfh@marvell.com>