arm64: fix for cortex-a53 erratum #845719
When running a compat (AArch32) userspace on Cortex-A53, a load at EL0
from a virtual address that matches the bottom 32 bits of the virtual
address used by a recent load at (AArch64) EL1 might return incorrect
data.

This patch fixes the issue by writing to the contextidr_el1
register on the exception return path when returning to a 32-bit task.

The patch has been backported from the ARM patch posted here:
http://permalink.gmane.org/gmane.linux.ports.arm.kernel/404196
http://lists.infradead.org/pipermail/linux-arm-kernel/2015-March/334664.html

The original patch uses the alternative runtime patching
introduced into arm64 in 3.18.
Here we just apply the fix at compile-time
when CONFIG_ARM64_ERRATUM_845719 is set.
This is ok for kernel image that only runs on affected A53.

Change-Id: If42aa2650bd8471a314796dd51492ab882232887
Signed-off-by: Anton Eidelman <antone@marvell.com>