char: vbus: add errors check after OTG plug out
when OTG device is attached it can draw current from
main battery which is driving 88pm886. For cases in which
the battery threshold is getting too low the otg_uvvbat
error will be flagged and OTG supply will be disabled.
in order to recover from this case we'll clear the errors
upon OTG device removal.

Change-Id: Ibcca4873d8d6185574d8b28ad21e1354fd0b5369
Signed-off-by: Amit Sharon <asharon@marvell.com>