power: 88pm88x_battery: always use SW battery detection
the battery detection circuit is monitoring the NTC pin and changes the BAT bit accordingly.
since the GPADC is sensitive, it can be affected by noises.
in case of a noise on the pin - the HW might think that battery was removed,
and then it will automatically set the CHG_EN bit.

in order to avoid such unexpected behavior,
disable the HW battery detection and always use manual method (SW)

Change-Id: If3faaa43451f6dc27f3d6a3a92b920f3f4a4da26
Signed-off-by: Shay Pathov <shayp@marvell.com>