mfd: mmp-map: clear array data befor use it
In reg set function, array index is used twice. If value length
is short while reg value is long, data after char "\n" is not 0.
When do kstrtouint, it will check the data after char "\n". If
it's not 0, it will return an error value and the operate will
be failed.

Change-Id: I7a3415e023364e443248600248b08600846c18ad
Signed-off-by: Leilei Shang <shangll@marvell.com>