mfd: 88pm88x-core: verify PMIC trimming
Added a function that verifies that the board's PMIC is trimmed
according to the Silicon design team.
If this is not the case it will print a warning.
Done for both 88pm880 and 88pm886.

Trimming is basically a calibration process done on every chip during the
manufacturing of the PMIC chips. The way to make sure a PMIC was trimmed is
to read the trim status registers (on test page):
pm880 - bits [4:0] of reg 0x92
pm886 - bit [7] of reg 0x8A

Change-Id: I958c7153dac762d2750e429463addf8ffe9763f9
Signed-off-by: Harel Manheim <hmanheim@marvell.com>