cpufreq: mmp: update the logic to update ddr up threshold
In previous implementation, when current freq
is >= swap_freq and goes to a higher one, the
ddr up threshold QoS will be canceled wrongly.
In this patch, we only consider the relationship
of current freq and swap_freq. When current freq
is >= swap_freq, send QoS request. When it becomes
lower then swap_freq, cancel QoS request.

Change-Id: I5fa9acdea4fc0cc83f112c9eb355551b7607c953
Signed-off-by: Bill Zhou <zhoub@marvell.com>