ASoC: mmp-tdm: fix audio record failed issue
For current setting, when only record is used, 4 slots
are set for tx. When record is shutdown, map will be
reset. But when free slot, tdm thinks tx still using
it, when next time record is open, tdm driver will not
configure tdm registers, so tdm is not working.

We should only allocate slots for tx when tx slot is set.

Change-Id: Ia8efb0fe7e9879cabb2255ef8c6411114d08478d
Signed-off-by: Leilei Shang <shangll@marvell.com>